requirements.txt
cachelib==0.9.0
certifi==2022.6.15
chardet==3.0.4
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
flask-mongoengine==1.0.0
Flask-Session==0.4.0
Flask-WTF==1.0.1
idna==2.7
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
pymongo==4.2.0
PyJWT==1.7.1
requests==2.20.0
urllib3==1.24.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
